server/pyproject.toml
[project]
name = "nl-it-translator"
version = "0.1.0"
description = "Dutch-Italian real-time voice translator with push-to-talk"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pipecat-ai[daily,webrtc,websocket,cartesia,silero,deepgram,openai,runner]>=1.0.0",
  "python-dotenv",
  "aiofiles"
]
